requirements.txt
langchain
langchain-openai
langchain-community
openai
fastapi
uvicorn
pydantic
python-dotenv
PyPDF2
requests
supabase
tiktoken
tavily-python
httpx
cachetools
